Proxy and Radio
A proxy server is an intermediate between the end device and another
server. The project is implemented with a simple socket programming in
C. This proxy server receives a HTTP request from the user agent (the
browser in this case) and parses the request. The request is parsed to
obtain the server URL. The corresponding IP address and the port number
of the web address is obtained. A new socket is created which is used to
connect to the destination server with this IP address. The HTTP request
from the browser is stored in the buffer and sent to the server with the
appropriate IP. The response is been sent back to the proxy which then
further sends it back to the browser.

A FTP proxy is an extension to the HTTP proxy which will also listen to
the FTP requests and serves the FTP responses along with the HTTP
requests and responses.

The UDP radio streams the live packet on to the network.